+++ History.md
@@ -1,3 +1,8 @@
+0.2.0 / 2015-04-11
+==================
+* Standalone socket server mode
+* Performance enhancement
+
 0.1.0 / 2015-03-13
 ==================
 * First Release